Render Pass generation is now working! (At least for some configurations)
Added:

-RenderPassCache Interface class.
-Validation asserts to CreateRenderPass() function.

Changed:

-Debug utils such as assertions are no longer compiled into release builds. Changes should be tested in debug builds before trying release builds.

Fixed:

-Warnings
-Release configuration is now compiling and running correctly.

+++ VKR/VKR/ExtensionManager.cpp
@@ -216,7 +216,7 @@ namespace PB
 
 	DynamicArray<const char*> ExtensionManager::GetEnabledLayers()
 	{
-		DynamicArray<const char*> output(m_availableLayers.size());
+		DynamicArray<const char*> output(static_cast<uint32_t>(m_availableLayers.size()));
 
 		for (auto& ext : m_availableLayers)
 		{
